TST: collect/share Index tests
