+++ pandas/tests/indexes/timedeltas/test_indexing.py
@@ -21,14 +21,6 @@ import pandas._testing as tm
 
 
 class TestGetItem:
-    def test_ellipsis(self):
-        # GH#21282
-        idx = timedelta_range("1 day", "31 day", freq="D", name="idx")
-
-        result = idx[...]
-        assert result.equals(idx)
-        assert result is not idx
-
     def test_getitem_slice_keeps_name(self):
         # GH#4226
         tdi = timedelta_range("1d", "5d", freq="H", name="timebucket")
