+++ pandas/tests/indexing/test_datetime.py
@@ -130,7 +130,7 @@ class TestDatetimeIndex:
         expected = DataFrame(-1, index=index, columns=["a"])
         tm.assert_frame_equal(result, expected)
 
-    def test_getitem_millisecond_resolution(self, frame_or_series):
+    def test_getitem_str_slice_millisecond_resolution(self, frame_or_series):
         # GH#33589
 
         keys = [
@@ -152,16 +152,3 @@ class TestDatetimeIndex:
             ],
         )
         tm.assert_equal(result, expected)
-
-    def test_str_subclass(self):
-        # GH 37366
-        class mystring(str):
-            pass
-
-        data = ["2020-10-22 01:21:00+00:00"]
-        index = pd.DatetimeIndex(data)
-        df = DataFrame({"a": [1]}, index=index)
-        df["b"] = 2
-        df[mystring("c")] = 3
-        expected = DataFrame({"a": [1], "b": [2], mystring("c"): [3]}, index=index)
-        tm.assert_equal(df, expected)
