+++ pandas/tests/indexing/test_scalar.py
@@ -77,6 +77,13 @@ class TestScalar(Base):
 class TestAtAndiAT:
     # at and iat tests that don't need Base class
 
+    def test_float_index_at_iat(self):
+        ser = Series([1, 2, 3], index=[0.1, 0.2, 0.3])
+        for el, item in ser.items():
+            assert ser.at[el] == item
+        for i in range(len(ser)):
+            assert ser.iat[i] == i + 1
+
     def test_at_iat_coercion(self):
 
         # as timestamp is not a tuple!
